% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calcNormFactors}
\alias{calcNormFactors}
\title{Cumulative sum scaling normalization factors

Return a vector of the the sum up to and including a quantile.}
\usage{
calcNormFactors(obj, p = cumNormStatFast(obj))
}
\arguments{
\item{obj}{An MRexperiment object or matrix.}

\item{p}{The pth quantile.}
}
\value{
Vector of the sum up to and including a sample's pth quantile.
}
\description{
Cumulative sum scaling normalization factors

Return a vector of the the sum up to and including a quantile.
}
\examples{
data(mouseData)
head(calcNormFactors(mouseData))
}
\seealso{
\code{\link{fitZig}} \code{\link{cumNormStatFast}} \code{\link{cumNorm}}
}

