% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitZeroLogNormal.R
\name{calcStandardError}
\alias{calcStandardError}
\title{Calculate the zero-inflated log-normal statistic's standard error}
\usage{
calcStandardError(mod, fitln, fitzero, coef = 2, exclude = NULL)
}
\arguments{
\item{mod}{The zero component model matrix}

\item{fitln}{A matrix with parameters from the log-normal fit}

\item{fitzero}{A matrix with parameters from the logistic fit}

\item{coef}{Coefficient of interest}

\item{exclude}{List of features to exclude}
}
\description{
Calculat the se for the model. Code modified from
"Adjusting for covariates in zero-inflated gamma and 
zero-inflated log-normal models for semicontinuous data", ED Mills
}
\seealso{
\code{\link{fitZeroLogNormal}} \code{\link{fitFeatureModel}}
}

