% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateEffectiveSamples.R
\name{calculateEffectiveSamples}
\alias{calculateEffectiveSamples}
\title{Estimated effective samples per feature}
\usage{
calculateEffectiveSamples(obj)
}
\arguments{
\item{obj}{The output of fitZig run on a MRexperiment object.}
}
\value{
A list of the estimated effective samples per feature.
}
\description{
Calculates the number of estimated effective samples per feature from the output
of a fitZig run. The estimated effective samples per feature is calculated as the
sum_1^n (n = number of samples) 1-z_i where z_i is the posterior probability a feature
belongs to the technical distribution.
}
\seealso{
\code{\link{fitZig}} \code{\link{MRcoefs}} \code{\link{MRfulltable}}
}

