% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitZeroLogNormal.R
\name{calcZeroAdjustment}
\alias{calcZeroAdjustment}
\title{Calculate the zero-inflated component's adjustment factor}
\usage{
calcZeroAdjustment(fitln, fitzero, mod, coef, exclude = NULL)
}
\arguments{
\item{fitln}{A matrix with parameters from the log-normal fit}

\item{fitzero}{A matrix with parameters from the logistic fit}

\item{mod}{The zero component model matrix}

\item{coef}{Coefficient of interest}

\item{exclude}{List of features to exclude}
}
\description{
Calculate the log ratio of average marginal probabilities for each sample
having a positive count. This becomes the adjustment factor for the log 
fold change.
}
\seealso{
\code{\link{fitZeroLogNormal}} \code{\link{fitFeatureModel}}
}

