% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitZeroLogNormal.R
\name{calcZeroComponent}
\alias{calcZeroComponent}
\title{Zero component}
\usage{
calcZeroComponent(mat, mod, weights)
}
\arguments{
\item{mat}{A matrix of normalized counts}

\item{mod}{A model matrix}

\item{weights}{Weight matrix for samples and counts}
}
\description{
Fit the zero (logisitic) component
}
\seealso{
\code{\link{fitZeroLogNormal}} \code{\link{fitFeatureModel}}
}
