% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{expSummary}
\alias{expSummary}
\alias{expSummary,MRexperiment-method}
\title{Access MRexperiment object experiment data}
\usage{
expSummary(obj)
}
\arguments{
\item{obj}{a \code{MRexperiment} object.}
}
\description{
The expSummary vectors represent the column (sample specific) sums of
features, i.e. the total number of reads for a sample, libSize and also the
normalization factors, normFactor.
}
\examples{
data(mouseData)
expSummary(mouseData)
}
\author{
Joseph N. Paulson, jpaulson@umiacs.umd.edu
}

