% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{fitFeatureModelResults-class}
\alias{fitFeatureModelResults-class}
\title{Class "fitFeatureModelResults" -- a formal class for storing results from a fitFeatureModel call}
\description{
This class contains all of the same information expected from a fitFeatureModel call, 
but it is defined in the S4 style as opposed to being stored as a list.
}
\section{Slots}{

\describe{
\item{\code{call}}{the call made to fitFeatureModel}

\item{\code{fitZeroLogNormal}}{list of parameter estimates for the zero-inflated log normal model}

\item{\code{design}}{model matrix}

\item{\code{taxa}}{taxa names}

\item{\code{counts}}{count matrix}

\item{\code{pvalues}}{calculated p-values}

\item{\code{permuttedFits}}{permutted z-score estimates under the null}
}}

