% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fitTimeSeries}
\alias{fitTimeSeries}
\title{Discover differentially abundant time intervals using SS-Anova

Calculate time intervals of interest using SS-Anova fitted models. Fitting is performed
uses Smoothing Spline ANOVA (SS-Anova) to find interesting intervals of time.
Given observations at different time points for two groups, fitTimeSeries
calculates a  function that models the difference in abundance between two
groups across all time. Using permutations we estimate a null distribution
of areas for the time intervals of interest and report significant intervals of time.
Use of the function for analyses should cite:
"Finding regions of interest in high throughput genomics data using smoothing splines"
Talukder H, Paulson JN, Bravo HC. (Submitted)}
\usage{
fitTimeSeries(obj, feature, class, time, id, lvl = NULL, B = 1000,
  seed = 123, norm = TRUE, sl = 1000)
}
\arguments{
\item{obj}{metagenomeSeq MRexperiment-class object.}

\item{feature}{Name or row of feature of interest.}

\item{class}{Name of column in phenoData of MRexperiment-class object for class memberhip.}

\item{time}{Name of column in phenoData of MRexperiment-class object for relative time.}

\item{id}{Name of column in phenoData of MRexperiment-class object for sample id.}

\item{lvl}{Vector or name of column in featureData of MRexperiment-class object for aggregating counts (if not OTU level).}

\item{B}{Number of permutations to perform}

\item{seed}{Random-number seed.}

\item{norm}{When aggregating counts to normalize or not.}

\item{sl}{Scaling value.}
}
\value{
Matrix of time point intervals of interest, Difference in abundance area and p-value.
}
\description{
Discover differentially abundant time intervals using SS-Anova

Calculate time intervals of interest using SS-Anova fitted models. Fitting is performed
uses Smoothing Spline ANOVA (SS-Anova) to find interesting intervals of time.
Given observations at different time points for two groups, fitTimeSeries
calculates a  function that models the difference in abundance between two
groups across all time. Using permutations we estimate a null distribution
of areas for the time intervals of interest and report significant intervals of time.
Use of the function for analyses should cite:
"Finding regions of interest in high throughput genomics data using smoothing splines"
Talukder H, Paulson JN, Bravo HC. (Submitted)
}
\examples{
data(mouseData)
fitTimeSeries(obj=mouseData,feature="Actinobacteria",
   class="status",id="mouseID",time="relativeTime",lvl='class',B=10)
}

