% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPi.R
\name{getPi}
\alias{getPi}
\title{Calculate the mixture proportions from the zero model / spike mass model
residuals.}
\usage{
getPi(residuals)
}
\arguments{
\item{residuals}{Residuals from the zero model.}
}
\value{
Mixture proportions for each sample.
}
\description{
F(x) = 1 / (1 + exp(-(x-m)/s)) (the CDF of the logistic distribution).
Provides the probability that a real-valued random variable X with a given
probability distribution will be found at a value less than or equal to x.
The output are the mixture proportions for the samples given the residuals
from the zero model.
}
\seealso{
\code{\link{fitZig}}
}
