% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{methods}
\name{libSize}
\alias{libSize}
\alias{libSize,MRexperiment-method}
\title{Access sample depth of coverage from MRexperiment object}
\usage{
libSize(obj)
}
\arguments{
\item{obj}{a \code{MRexperiment} object.}
}
\value{
Library sizes
}
\description{
The libSize vector represents the column (sample specific) sums of features,
i.e. the total number of reads for a sample or depth of coverage. It is used by
\code{\link{fitZig}}.
}
\examples{

data(lungData)
head(libSize(lungData))

}
\author{
Joseph N. Paulson, jpaulson@umiacs.umd.edu
}

