% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_biom.R
\name{load_biom}
\alias{load_biom}
\title{Load objects organized in the Biom format.}
\usage{
load_biom(file)
}
\arguments{
\item{file}{The biom object filepath.}
}
\value{
A MRexperiment object.
}
\description{
Wrapper to load Biom formatted object.
}
\examples{
#library(biomformat)
#rich_dense_file = system.file("extdata", "rich_dense_otu_table.biom", package = "biomformat")
#x = load_biom(rich_dense_file)
#x
}
\seealso{
\code{\link{load_meta}} \code{\link{load_phenoData}} \code{\link{newMRexperiment}} \code{\link{biom2MRexperiment}}
}

