% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_meta.R
\name{load_meta}
\alias{load_meta}
\alias{metagenomicLoader}
\title{Load a count dataset associated with a study.}
\usage{
load_meta(file, sep = "\\t")
}
\arguments{
\item{file}{Path and filename of the actual data file.}

\item{sep}{File delimiter.}
}
\value{
A list with objects 'counts' and 'taxa'.
}
\description{
Load a matrix of OTUs in a tab delimited format
}
\examples{
dataDirectory <- system.file("extdata", package="metagenomeSeq")
lung = load_meta(file.path(dataDirectory,"CHK_NAME.otus.count.csv"))
}
\seealso{
\code{\link{load_phenoData}}
}

