% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadBiom.R
\name{loadBiom}
\alias{loadBiom}
\title{Load objects organized in the Biom format.}
\usage{
loadBiom(file)
}
\arguments{
\item{file}{The biom object filepath.}
}
\value{
A MRexperiment object.
}
\description{
Wrapper to load Biom formatted object.
}
\examples{

#library(biomformat)
#rich_dense_file = system.file("extdata", "rich_dense_otu_table.biom", package = "biomformat")
#x = loadBiom(rich_dense_file)
#x
}
\seealso{
\code{\link{loadMeta}} \code{\link{loadPhenoData}} \code{\link{newMRexperiment}} \code{\link{biom2MRexperiment}}
}

