% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMeta.R
\name{loadMeta}
\alias{loadMeta}
\alias{metagenomicLoader}
\title{Load a count dataset associated with a study.}
\usage{
loadMeta(file, sep = "\\t")
}
\arguments{
\item{file}{Path and filename of the actual data file.}

\item{sep}{File delimiter.}
}
\value{
A list with objects 'counts' and 'taxa'.
}
\description{
Load a matrix of OTUs in a tab delimited format
}
\examples{

dataDirectory <- system.file("extdata", package="metagenomeSeq")
lung = loadMeta(file.path(dataDirectory,"CHK_NAME.otus.count.csv"))

}
\seealso{
\code{\link{loadPhenoData}}
}
