% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_metagenomeSeq_function.R
\docType{package}
\name{metagenomeSeq-deprecated}
\alias{metagenomeSeq-deprecated}
\alias{deprecated_metagenomeSeq_function}
\alias{fitMeta}
\alias{load_phenoData}
\alias{load_meta}
\alias{load_biom}
\alias{load_metaQ}
\title{Depcrecated functions in the metagenomeSeq package.}
\usage{
deprecated_metagenomeSeq_function(x, value, ...)
}
\arguments{
\item{x}{For assignment operators, the object that will undergo a replacement
(object inside parenthesis).}

\item{value}{For assignment operators, the value to replace with 
(the right side of the assignment).}

\item{...}{For functions other than assignment operators, 
parameters to be passed to the modern version of the function (see table).}
}
\description{
These functions may be removed completely in the next release.
}
