\name{metagenomeSeq-package}
\docType{package}
\alias{metagenomeSeq}
\alias{metagenomeSeq-package}
\title{Statistical analysis for sparse high-throughput sequencing}
\description{
metagenomeSeq is designed to determine features (be it Operational Taxanomic Unit (OTU), species, etc.) that are differentially abundant between two or more groups of multiple samples. metagenomeSeq is designed to address the effects of both normalization and under-sampling of microbial communities on disease association detection and the testing of feature correlations.

A user's guide is available, and can be opened by typing \code{vignette("metagenomeSeq")} 

The metagenomeSeq package implements novel normalization and statistical methodology in the following papers.
}

\author{
Paulson, JN <jpaulson@umiacs.umd.edu>; Pop, M; Corrada Bravo, H
}

\references{
Paulson, Joseph N., O. Colin Stine, Héctor Corrada Bravo, and Mihai Pop. "Differential   abundance analysis for microbial marker-gene surveys." Nature methods (2013).
}
\keyword{package}
