\name{mouseData}
\docType{data}
\alias{mouseData}

\title{OTU abundance matrix of mice samples from a diet longitudinal study}
\description{This is a list with a matrix of OTU counts, taxa annotations for each OTU, otu names, and vector of phenotypic data. Samples along the columns and OTUs along the rows.}
\value{
MRexperiment-class object of 16S mouse samples.
}
\usage{mouseData}
\format{A list of OTU matrix, taxa, otus, and phenotypes}
\references{http://www.ncbi.nlm.nih.gov/pmc/articles/PMC2894525/}