% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MRexperiment2biom.R
\name{MRexperiment2biom}
\alias{MRexperiment2biom}
\title{MRexperiment to biom objects}
\usage{
MRexperiment2biom(obj, id = NULL, norm = FALSE, log = FALSE, sl = 1000,
  qiimeVersion = TRUE)
}
\arguments{
\item{obj}{The MRexperiment object.}

\item{id}{Optional id for the biom matrix.}

\item{norm}{Normalized data?}

\item{log}{Logged data?}

\item{sl}{scaling factor for normalized counts.}

\item{qiimeVersion}{Format fData according to QIIME specifications (assumes only taxonomy in fData).}
}
\value{
A biom object.
}
\description{
Wrapper to convert MRexperiment objects to biom objects.
}
\seealso{
\code{\link{load_meta}} \code{\link{load_phenoData}} \code{\link{newMRexperiment}} \code{\link{load_biom}} \code{\link{biom2MRexperiment}}
}

