% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRihw.R
\name{MRihw}
\alias{MRihw}
\title{MRihw runs IHW within a MRcoefs() call}
\usage{
MRihw(obj, ...)
}
\arguments{
\item{obj}{Either a fitFeatureModelResults or fitZigResults object}

\item{...}{other parameters}
}
\description{
Function used in MRcoefs() when "IHW" is set as the p value adjustment method
}
