% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{methods}
\name{normFactors<-}
\alias{normFactors<-}
\alias{normFactors<-,MRexperiment,numeric-method}
\title{Replace the normalization factors in a MRexperiment object}
\usage{
\S4method{normFactors}{MRexperiment,numeric}(object) <- value
}
\arguments{
\item{object}{a \code{MRexperiment} object}

\item{value}{vector of normalization scaling factors}
}
\value{
Normalization scaling factors
}
\description{
Function to replace the scaling factors, aka the normalization factors, of
samples in a MRexperiment object.
}
\examples{

data(lungData)
head(normFactors(lungData)<- rnorm(1))

}
\author{
Joseph N. Paulson
}
