% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotClassTimeSeries}
\alias{plotClassTimeSeries}
\title{Plot abundances by class}
\usage{
plotClassTimeSeries(res, xlab = "Time", ylab = "Abundance",
  color0 = "black", color1 = "red", ...)
}
\arguments{
\item{res}{Output of fitTimeSeries function}

\item{xlab}{X-label.}

\item{ylab}{Y-label.}

\item{color0}{Color of samples from first group.}

\item{color1}{Color of samples from second group.}

\item{...}{Extra plotting arguments.}
}
\description{
Plot abundances by class
}
\details{
Plot the abundance of values for each class using
a spline approach on the estimated full model.
}
\examples{
data(mouseData)
res = fitTimeSeries(obj=mouseData,feature="Actinobacteria",
   class="status",id="mouseID",time="relativeTime",lvl='class',B=10)
plotClassTimeSeries(res,pch=21,bg=res$data$class,ylim=c(0,8))
}
\seealso{
\code{\link{fitTimeSeries}}
}

