\name{posterior.probs}
\Rdversion{1.0}
\docType{methods}
\alias{posterior.probs,MRexperiment-method}
\alias{posterior.probs}
\title{
   Access the posterior probabilities that results from analysis
}

\description{
   Accessing the posterior probabilities following a run through \code{\link{fitZig}}
}

\usage{
\S4method{posterior.probs}{MRexperiment}(obj)
}

\arguments{
  \item{obj}{a \code{MRexperiment} object.}
}

\author{
   Joseph N. Paulson, jpaulson@umiacs.umd.edu
}
\examples{
# see vignette
}   
   
