% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allClasses.R
\name{returnAppropriateObj}
\alias{returnAppropriateObj}
\title{Check if MRexperiment or matrix and return matrix}
\usage{
returnAppropriateObj(obj, norm, log, sl = 1000)
}
\arguments{
\item{obj}{a \code{MRexperiment} or \code{matrix} object}

\item{norm}{return a normalized \code{MRexperiment} matrix}

\item{log}{return a log transformed \code{MRexperiment} matrix}

\item{sl}{scaling value}
}
\value{
Matrix
}
\description{
Function to check if object is a MRexperiment
class or matrix
}
\examples{
data(lungData)
head(returnAppropriateObj(lungData,norm=FALSE,log=FALSE))
}

