% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkZig.R
\name{shrinkZig}
\alias{shrinkZig}
\title{Shrink coefficients of a fitZig result}
\usage{
shrinkZig(obj, coef = 2)
}
\arguments{
\item{obj}{result of call to `fitZig` or `.do_fitZig` functions}

\item{coef}{coefficient to shrink}
}
\description{
Shrink coefficients of a fitZig result
}
