% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fitTimeSeries.R
\name{ssIntervalCandidate}
\alias{ssIntervalCandidate}
\title{calculate interesting time intervals

Calculates time intervals of interest using SS-Anova fitted confidence intervals.}
\usage{
ssIntervalCandidate(fit, standardError, timePoints, positive = TRUE, C = 0)
}
\arguments{
\item{fit}{SS-Anova fits.}

\item{standardError}{SS-Anova se estimates.}

\item{timePoints}{Time points interpolated over.}

\item{positive}{Positive region or negative region (difference in abundance is positive/negative).}

\item{C}{Value for which difference function has to be larger or smaller than (default 0).}
}
\value{
Matrix of time point intervals of interest
}
\description{
calculate interesting time intervals

Calculates time intervals of interest using SS-Anova fitted confidence intervals.
}
\examples{
# Not run
}
\seealso{
\code{\link{cumNorm}} \code{\link{fitTimeSeries}} \code{\link{ssFit}} \code{\link{ssPerm}} \code{\link{ssPermAnalysis}}
}

