% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ssPermAnalysis}
\alias{ssPermAnalysis}
\title{smoothing-splines anova fits for each permutation

Calculates the fit for each permutation and estimates the area under the null (permutted)
model for interesting time intervals of differential abundance.}
\usage{
ssPermAnalysis(data, permList, intTimes, timePoints)
}
\arguments{
\item{data}{Data used in estimation.}

\item{permList}{A list of permutted class memberships}

\item{intTimes}{Interesting time intervals.}

\item{timePoints}{Time points to interpolate over.}
}
\value{
A matrix of permutted area estimates for time intervals of interest.
}
\description{
smoothing-splines anova fits for each permutation

Calculates the fit for each permutation and estimates the area under the null (permutted)
model for interesting time intervals of differential abundance.
}
\examples{
# Not run
}

