% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumNorm.R
\name{calcNormFactors}
\alias{calcNormFactors}
\title{Cumulative sum scaling (css) normalization factors}
\usage{
calcNormFactors(obj, p = cumNormStatFast(obj))
}
\arguments{
\item{obj}{An MRexperiment object or matrix.}

\item{p}{The pth quantile.}
}
\value{
Vector of the sum up to and including a sample's pth quantile.
}
\description{
Return a vector of the the sum up to and including a quantile.
}
\examples{

data(mouseData)
head(calcNormFactors(mouseData))

}
\seealso{
\code{\link{fitZig}} \code{\link{cumNormStatFast}} \code{\link{cumNorm}}
}
