% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{methods}
\name{expSummary}
\alias{expSummary}
\alias{expSummary,MRexperiment-method}
\title{Access MRexperiment object experiment data}
\usage{
expSummary(obj)
}
\arguments{
\item{obj}{a \code{MRexperiment} object.}
}
\value{
Experiment summary table
}
\description{
The expSummary vectors represent the column (sample specific) sums of
features, i.e. the total number of reads for a sample, libSize and also the
normalization factors, normFactor.
}
\examples{

data(mouseData)
expSummary(mouseData)

}
\author{
Joseph N. Paulson, jpaulson@umiacs.umd.edu
}

