\name{lungData}
\docType{data}
\alias{lungData}

\title{OTU abundance matrix of samples from a smoker/non-smoker study}
\description{This is a list with a matrix of OTU counts,otu names, taxa annotations for each OTU, and phenotypic data. Samples along the columns and OTUs along the rows.}
\value{
MRexperiment-class object of 16S lung samples.
}
\usage{lungData}
\format{A list of OTU matrix, taxa, otus, and phenotypes}
\references{http://www.ncbi.nlm.nih.gov/pubmed/21680950}