% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRihw.R
\docType{methods}
\name{MRihw,fitZigResults-method}
\alias{MRihw,fitZigResults-method}
\title{MRihw runs IHW within a MRcoefs() call}
\usage{
\S4method{MRihw}{fitZigResults}(obj, p, adjustMethod, alpha)
}
\arguments{
\item{obj}{Either a fitFeatureModelResults or fitZigResults object}

\item{p}{a vector of pvalues extracted from obj}

\item{alpha}{pvalue significance level specified for IHW call. Default is 0.1}

\item{IHWcov}{Value specifying which covariate to use in IHW pvalue adjustment. For obj
of class \code{\link{fitFeatureModelResults}}, options are "median" and "nnz" (number of non-zero 
features per row). For obj of class \code{\link{fitZigResults}}, options are "Amean" and "nnz".}
}
\description{
Function used in MRcoefs() when "IHW" is set as the p value adjustment method
}
