% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{normFactors}
\alias{normFactors}
\alias{normFactors,MRexperiment-method}
\title{Access the normalization factors in a MRexperiment object}
\usage{
normFactors(obj)
}
\arguments{
\item{obj}{a \code{MRexperiment} object.}
}
\description{
Function to access the scaling factors, aka the normalization factors, of
samples in a MRexperiment object.
}
\examples{
data(lungData)
head(normFactors(lungData))
}
\author{
Joseph N. Paulson, jpaulson@umiacs.umd.edu
}

