% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{methods}
\name{normFactors}
\alias{normFactors}
\title{Access the normalization factors in a MRexperiment object}
\usage{
normFactors(object)
}
\arguments{
\item{object}{a \code{MRexperiment} object}
}
\value{
Normalization scaling factors
}
\description{
Function to access the scaling factors, aka the normalization factors, of
samples in a MRexperiment object.
}
\examples{

data(lungData)
head(normFactors(lungData))

}
\author{
Joseph N. Paulson
}
