% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{posterior.probs}
\alias{posterior.probs}
\alias{posterior.probs,MRexperiment-method}
\title{Access the posterior probabilities that results from analysis}
\usage{
posterior.probs(obj)
}
\arguments{
\item{obj}{a \code{MRexperiment} object.}
}
\description{
Accessing the posterior probabilities following a run through
\code{\link{fitZig}}
}
\examples{
# see vignette
}
\author{
Joseph N. Paulson, jpaulson@umiacs.umd.edu
}

