% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ssFit}
\alias{ssFit}
\title{smoothing-splines anova fit}
\usage{
ssFit(abundance, class, time, id, ...)
}
\arguments{
\item{abundance}{Numeric vector of abundances.}

\item{class}{Class membership (factor of group membership).}

\item{time}{Time point vector of relative times (same length as abundance).}

\item{id}{Sample / patient id.}

\item{...}{Extra parameters for ssanova function (see ?ssanova).}
}
\value{
A list containing:
     data        : Inputed data
     fit         : The interpolated / fitted values for timePoints
     se          : The standard error for CI intervals
     timePoints  : The time points interpolated over
}
\description{
smoothing-splines anova fit
}
\details{
Sets up a data-frame with the feature abundance,
class information, time points, sample ids and returns
the fitted values for the fitted model.
}
\examples{
# Not run
}

