% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ssPerm}
\alias{ssPerm}
\title{class permutations for smoothing-spline time series analysis

Creates a list of permuted class memberships for the time series permuation tests.}
\usage{
ssPerm(df, B)
}
\arguments{
\item{df}{Data frame containing class membership and sample/patient id label.}

\item{B}{Number of permutations.}
}
\value{
A list of permutted class memberships
}
\description{
class permutations for smoothing-spline time series analysis

Creates a list of permuted class memberships for the time series permuation tests.
}
\examples{
# Not run
}
\seealso{
\code{\link{cumNorm}} \code{\link{fitTimeSeries}} \code{\link{ssFit}} \code{\link{ssPermAnalysis}} \code{\link{ssIntervalCandidate}}
}

