% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ssPermAnalysis}
\alias{ssPermAnalysis}
\title{smoothing-splines anova fits for each permutation}
\usage{
ssPermAnalysis(data, permList, intTimes, timePoints, ...)
}
\arguments{
\item{data}{Data used in estimation.}

\item{permList}{A list of permutted class memberships}

\item{intTimes}{Interesting time intervals.}

\item{timePoints}{Time points to interpolate over.}

\item{...}{Options for ssanova}
}
\value{
A matrix of permutted area estimates for time intervals of interest.
}
\description{
smoothing-splines anova fits for each permutation
}
\details{
Calculates the fit for each permutation and estimates
the area under the null (permutted) model for interesting time
intervals of differential abundance.
}
\examples{
# Not run
}
\seealso{
\code{\link{cumNorm}} \code{\link{fitTimeSeries}} \code{\link{ssFit}} \code{\link{ssPerm}} \code{\link{ssIntervalCandidate}}
}

